requirements.txt
# AI Avatar Project Dependencies
# Core dependencies for Azure AI Services Text-to-Speech Avatar with Flask

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Azure AI Services
azure-cognitiveservices-speech==1.34.1
azure-core==1.29.6
azure-identity==1.15.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-resource==23.1.1
azure-cli-core==2.56.0

# Additional Azure SDKs for comprehensive support
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0

# HTTP and API handling
requests==2.31.0
urllib3==2.1.0

# Environment and configuration
python-dotenv==1.0.0
python-decouple==3.8

# Audio processing (for AI Avatar)
pydub==0.25.1
numpy==1.24.4

# JSON handling and validation
jsonschema==4.20.0
pydantic==2.5.2

# Logging and utilities
loguru==0.7.2
click==8.1.7

# Date and time utilities
python-dateutil==2.8.2

# WebSocket support (if needed for real-time features)
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Base64 and encoding utilities
base64io==1.0.3

# HTTP client with better error handling
httpx==0.25.2

# Async support
asyncio-mqtt==0.13.0
